Refactored the UI, added a web_playground.py to be run without any models.
